Add files via upload
The project is modularized, with separate classes for Ball, Paddle, Score, and GamePanel to improve readability and understanding. 
To experience the Pong Game, clone or download the repository, ensure all classes (Pong, GameFrame, GamePanel, Paddle, Ball, Score) are in the PONG_GAME package, and compile them together. Then, run the Pong class, which initializes the game window and starts the animation. Use the keyboard controls (1/2 for Player 1 and 9/0 for Player 2) to move paddles and enjoy the interactive gameplay.